# Add ruff parser backend as experimental alternative to tree-sitter