… pylight/Cargo.lock
… pylight/Cargo.toml
… pylight/static/devtools.html
… pylight/tests/integration/test_file_watcher.rs
… pylight/tests/integration/test_ignore_dirs.rs
… pylight/tests/integration/test_index.rs
… pylight/tests/integration/test_lsp.rs
… pylight/tests/parallel_indexing_test.rs