## Summary

This PR adds support for using ruff's Python parser as an experimental alternative to tree-sitter, while maintaining full backward compatibility. The tree-sitter parser remains the default.

## Changes

### Parser Abstraction
- Created a `Parser` trait to abstract symbol extraction logic
- Implemented `TreeSitterParser` wrapper for existing tree-sitter functionality
- Implemented `RuffParser` using ruff's AST for symbol extraction
- Both parsers extract identical symbols (only column positions differ slightly)

### CLI and Configuration
- Added `--parser` CLI flag to select backend (default: tree-sitter)
- Parser backend selection flows through the entire system (LSP server, index, updater)
- Implemented standard `FromStr` trait for `ParserBackend` enum

### DevTools Improvements
- Split UI into tabs: "Symbol Search" and "Parser Comparison"
- Added parser backend dropdown in Symbol Search tab
- Parser selection is passed to pylight when spawning LSP server
- Improved tab styling with blue active states and better contrast
- Removed rounded corners for cleaner, modern look

### Testing and Validation
- All existing tests updated to use `SymbolIndex::default()`
- Created test program to verify both parsers extract same symbols
- Added parser comparison tool in devtools for side-by-side results
- Verified on multiple Python codebases - both parsers find identical symbols

## Performance

Initial testing shows comparable performance between the two parsers. The ruff parser may have advantages for certain use cases, particularly when integrating with other ruff-based tooling.

## Future Work

- Add benchmarks comparing parser performance on large codebases
- Consider making ruff the default parser after more testing
- Integrate with ruff's error recovery for better handling of invalid syntax

## Test Plan

- [x] All existing tests pass
- [x] Manual testing with devtools shows identical symbol extraction
- [x] Parser selection works correctly in LSP mode
- [x] DevTools UI improvements tested in browser

🤖 Generated with [Claude Code](https://claude.ai/code)

Co-Authored-By: Claude <noreply@anthropic.com>